+++ src/modules/user/user.service.ts
@@ -1,10 +1,14 @@
-import { BadRequestException, Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { PrismaService } from 'src/common/prisma/prisma.service';
 import { JwtService } from '@nestjs/jwt';
 import { ConfigService } from '@nestjs/config';
-
+import * as bcrypt from 'bcrypt';
 @Injectable()
 export class UserService {
   constructor(
@@ -13,16 +17,34 @@ export class UserService {
     private readonly config: ConfigService,
   ) {}
 
-  create(createUserDto: CreateUserDto) {
-    const { email, ho_ten, mat_khau, so_dt, loai_nguoi_dung } = createUserDto;
+  async create(createUserDto: CreateUserDto) {
+    const { email, ho_ten, mat_khau, so_dt, ma_loai_nguoi_dung } =
+      createUserDto;
 
+    if (!email || !mat_khau || !ho_ten || !so_dt) {
+      // console.log({ email, mat_khau, ho_ten, so_dien_thoai });
+      throw new BadRequestException('Dữ liệu truyền vào không phù hợp!');
+    }
+
+    const userExist = await this.prisma.nguoiDung.findFirst({
+      where: {
+        email: email,
+      },
+    });
+
+    if (userExist)
+      throw new BadRequestException('Đã tồn tại người dùng có email này!');
+
+    // console.log(ma_loai_nguoi_dung);
+
+    const hashedPassword = await bcrypt.hash(mat_khau, 10);
     const newUser = this.prisma.nguoiDung.create({
       data: {
         email: email,
         ho_ten: ho_ten,
-        mat_khau: mat_khau,
+        mat_khau: hashedPassword,
         so_dt: so_dt,
-        loai_nguoi_dung: loai_nguoi_dung,
+        ma_loai_nguoi_dung: ma_loai_nguoi_dung,
       },
     });
 
@@ -35,10 +57,21 @@ export class UserService {
       orderBy: {
         tai_khoan: 'asc',
       },
+      select: {
+        tai_khoan: true,
+        email: true,
+        ho_ten: true,
+        so_dt: true,
+        LoaiNguoiDung: true,
+      },
     });
+
+    if (!danhSachNguoiDung)
+      throw new NotFoundException('Không tìm thấy thông tin!');
+
     return {
       totalItems: totalItems,
-      danhSachNguoiDung: danhSachNguoiDung || [],
+      danhSachNguoiDung: danhSachNguoiDung,
     };
   }
 
@@ -52,36 +85,51 @@ export class UserService {
       take: pageSize,
       skip: skip,
       orderBy: { tai_khoan: 'asc' },
+      select: {
+        tai_khoan: true,
+        email: true,
+        ho_ten: true,
+        so_dt: true,
+        LoaiNguoiDung: true,
+      },
     });
+
+    if (!result) throw new NotFoundException('Không tìm thấy thông tin!');
+
     return {
       page,
       pageSize,
       totalPages: totalPages,
       totalItems: totalItems,
-      items: result || [],
+      items: result,
     };
   }
 
   async findUserType() {
-    const danhSachLoaiNguoiDung = await this.prisma.nguoiDung.findMany({
-      select: {
-        loai_nguoi_dung: true,
-      },
-    });
+    const danhSachLoaiNguoiDung = await this.prisma.loaiNguoiDung.findMany();
     return danhSachLoaiNguoiDung;
   }
 
   async searchUser(hoTen: string) {
     const foundUsers = await this.prisma.nguoiDung.findMany({
       where: {
-        ho_ten: hoTen,
+        ho_ten: { contains: hoTen },
       },
       orderBy: {
         tai_khoan: 'asc',
       },
+      select: {
+        tai_khoan: true,
+        email: true,
+        ho_ten: true,
+        so_dt: true,
+        LoaiNguoiDung: true,
+      },
     });
 
-    return { danhSachNguoiDung: foundUsers || [] };
+    if (!foundUsers) throw new NotFoundException('Không tìm thấy thông tin!');
+
+    return { danhSachNguoiDung: foundUsers };
   }
 
   async searchUserPagination(hoTen: string, page: number, pageSize: number) {
@@ -92,33 +140,55 @@ export class UserService {
     const totalPages = Math.ceil(totalItems / pageSize);
     const result = await this.prisma.nguoiDung.findMany({
       where: {
-        ho_ten: hoTen,
+        ho_ten: { contains: hoTen },
       },
       take: pageSize,
       skip: skip,
       orderBy: { tai_khoan: 'asc' },
+      select: {
+        tai_khoan: true,
+        email: true,
+        ho_ten: true,
+        so_dt: true,
+        LoaiNguoiDung: true,
+      },
     });
+
+    if (!result) throw new NotFoundException('Không tìm thấy thông tin!');
     return {
       page,
       pageSize,
       totalPages: totalPages,
       totalItems: totalItems,
-      items: result || [],
+      items: result,
     };
   }
 
   async userInfo(token: string) {
+    // console.log(token);
+    if (!token) throw new BadRequestException('Vui lòng nhập token!');
+
     const decoded = this.jwt.verify(token, {
-      secret: this.config.get<string>('ACCESS_TOKEN_SECRET'),
+      secret: this.config.get<string>('REFRESH_TOKEN_SECRET'),
     });
 
-    if (!decoded) throw new BadRequestException('Token sai!');
+    if (!decoded || !decoded.nguoiDung)
+      throw new BadRequestException('Token sai!');
+
+    console.log(decoded);
 
     const user = await this.prisma.nguoiDung.findUnique({
-      where: { tai_khoan: decoded.tai_khoan },
+      where: { tai_khoan: decoded.nguoiDung },
+      select: {
+        tai_khoan: true,
+        email: true,
+        ho_ten: true,
+        so_dt: true,
+        LoaiNguoiDung: true,
+      },
     });
 
-    if (!user) throw new BadRequestException('Không tìm thấy người dùng!');
+    if (!user) throw new NotFoundException('Không tìm thấy người dùng!');
 
     return user;
   }
@@ -128,32 +198,77 @@ export class UserService {
       where: {
         tai_khoan: +taiKhoan,
       },
+      select: {
+        tai_khoan: true,
+        email: true,
+        ho_ten: true,
+        so_dt: true,
+        LoaiNguoiDung: true,
+      },
     });
 
-    if (!user) throw new BadRequestException('Không tìm thấy người dùng!');
+    if (!user) throw new NotFoundException('Không tìm thấy người dùng!');
 
     return user;
   }
 
-  update(id: number, updateUserDto: UpdateUserDto) {
-    const { email, ho_ten, mat_khau, so_dt, loai_nguoi_dung } = updateUserDto;
+  async update(id: number, updateUserDto: UpdateUserDto) {
+    const foundUser = await this.prisma.nguoiDung.findUnique({
+      where: { tai_khoan: +id },
+    });
+
+    if (!foundUser)
+      throw new BadRequestException('Không tồn tại người dùng này!');
+
+    const { email, ho_ten, mat_khau, so_dt, ma_loai_nguoi_dung } =
+      updateUserDto;
+    if (!email || !mat_khau || !ho_ten || !so_dt) {
+      // console.log({ email, mat_khau, ho_ten, so_dien_thoai });
+      throw new BadRequestException('Dữ liệu truyền vào không phù hợp!');
+    }
+    const hashedPassword = await bcrypt.hash(mat_khau, 10);
     const updateUser = this.prisma.nguoiDung.update({
-      where: { tai_khoan: id },
+      where: { tai_khoan: +id },
       data: {
         ho_ten: ho_ten,
         email: email,
-        mat_khau: mat_khau,
+        mat_khau: hashedPassword,
         so_dt: so_dt,
-        loai_nguoi_dung: loai_nguoi_dung,
+        ma_loai_nguoi_dung: ma_loai_nguoi_dung,
+      },
+      select: {
+        tai_khoan: true,
+        email: true,
+        ho_ten: true,
+        so_dt: true,
+        LoaiNguoiDung: true,
       },
     });
 
     return updateUser;
   }
 
-  remove(id: number) {
-    const deleteUser = this.prisma.nguoiDung.delete({
-      where: { tai_khoan: id },
+  async remove(id: number) {
+    const foundUser = await this.prisma.nguoiDung.findUnique({
+      where: { tai_khoan: +id },
+    });
+
+    if (!foundUser)
+      throw new NotFoundException('Không tồn tại người dùng này!');
+
+    await this.prisma.datVe.deleteMany({
+      where: { tai_khoan: +id },
+    });
+
+    const deleteUser = await this.prisma.nguoiDung.delete({
+      where: { tai_khoan: +id },
+      select: {
+        tai_khoan: true,
+        email: true,
+        ho_ten: true,
+        so_dt: true,
+        LoaiNguoiDung: true,
+      },
     });
 
     return deleteUser;
